+++ ParallaxGenLib/src/BethesdaDirectory.cpp
@@ -122,8 +122,8 @@ auto BethesdaDirectory::getFile(const filesystem::path &RelPath, const bool &Cac
     bsa::tes4::version BSAVersion = BSAStruct->Version;
     bsa::tes4::archive BSAObj = BSAStruct->Archive;
 
-    string ParentPath = wstringToString(RelPath.parent_path().wstring());
-    string Filename = wstringToString(RelPath.filename().wstring());
+    string ParentPath = wstrToStr(RelPath.parent_path().wstring());
+    string Filename = wstrToStr(RelPath.filename().wstring());
 
     const auto File = BSAObj[ParentPath][Filename];
     if (File) {
@@ -133,7 +133,7 @@ auto BethesdaDirectory::getFile(const filesystem::path &RelPath, const bool &Cac
         File->write(AOS, BSAVersion);
       } catch (const std::exception &E) {
         if (Logging) {
-          spdlog::error(L"Failed to read file {}: {}", RelPath.wstring(), stringToWstring(E.what()));
+          spdlog::error(L"Failed to read file {}: {}", RelPath.wstring(), strToWstr(E.what()));
         }
       }
 
@@ -279,7 +279,7 @@ void BethesdaDirectory::addBSAFilesToMap() {
       addBSAToFileMap(BSAName);
     } catch (const std::exception &E) {
       if (Logging) {
-        spdlog::error(L"Failed to add BSA file {} to map (Skipping): {}", BSAName, stringToWstring(E.what()));
+        spdlog::error(L"Failed to add BSA file {} to map (Skipping): {}", BSAName, strToWstr(E.what()));
       }
       continue;
     }
@@ -311,7 +311,7 @@ void BethesdaDirectory::addLooseFilesToMap() {
       }
     } catch (const std::exception &E) {
       if (Logging) {
-        spdlog::error(L"Failed to load file from iterator (Skipping): {}", stringToWstring(E.what()));
+        spdlog::error(L"Failed to load file from iterator (Skipping): {}", strToWstr(E.what()));
       }
       continue;
     }
@@ -371,7 +371,7 @@ void BethesdaDirectory::addBSAToFileMap(const wstring &BSAName) {
       }
     } catch (const std::exception &E) {
       if (Logging) {
-        spdlog::error(L"Failed to get file pointer from BSA, skipping {}: {}", BSAName, stringToWstring(E.what()));
+        spdlog::error(L"Failed to get file pointer from BSA, skipping {}: {}", BSAName, strToWstr(E.what()));
       }
       continue;
     }
@@ -488,10 +488,10 @@ auto BethesdaDirectory::getBSAFilesFromINIs() const -> vector<wstring> {
     // loop through each ini file
     wstring INIVal;
     for (const auto &INIPath : INIFileOrder) {
-      wstring CurVal = readINIValue(INIPath, L"Archive", stringToWstring(Field), Logging, FirstINIRead);
+      wstring CurVal = readINIValue(INIPath, L"Archive", strToWstr(Field), Logging, FirstINIRead);
 
       if (Logging) {
-        spdlog::trace(L"Found ini key pair from INI {}: {}: {}", INIPath.wstring(), stringToWstring(Field), CurVal);
+        spdlog::trace(L"Found ini key pair from INI {}: {}: {}", INIPath.wstring(), strToWstr(Field), CurVal);
       }
 
       if (CurVal.empty()) {
@@ -508,7 +508,7 @@ auto BethesdaDirectory::getBSAFilesFromINIs() const -> vector<wstring> {
     }
 
     if (Logging) {
-      spdlog::trace(L"Found BSA files from INI field {}: {}", stringToWstring(Field), INIVal);
+      spdlog::trace(L"Found BSA files from INI field {}: {}", strToWstr(Field), INIVal);
     }
 
     // split into components