+++ ParallaxGenLib/src/NIFUtil.cpp
@@ -270,7 +270,7 @@ auto NIFUtil::configureShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP,
 
 // Texture slot helpers
 auto NIFUtil::setTextureSlot(nifly::NifFile *NIF, nifly::NiShape *NIFShape, const TextureSlots &Slot, const std::wstring &TexturePath, bool &Changed) -> void {
-  auto TexturePathStr = ParallaxGenUtil::wstringToString(TexturePath);
+  auto TexturePathStr = ParallaxGenUtil::wstrToStr(TexturePath);
   setTextureSlot(NIF, NIFShape, Slot, TexturePathStr, Changed);
 }