+++ ParallaxGenLib/src/patchers/PatcherVanillaParallax.cpp
@@ -113,7 +113,7 @@ auto PatcherVanillaParallax::shouldApply(NiShape *NIFShape, const array<wstring,
   if (!DiffuseMap.empty() && !PGD->isFile(DiffuseMap)) {
     // no Diffuse map
     spdlog::trace(L"NIF: {} | Shape: {} | Parallax | Shape Rejected: Diffuse map missing: {}", NIFPath.wstring(),
-                  ShapeBlockID, stringToWstring(DiffuseMap));
+                  ShapeBlockID, strToWstr(DiffuseMap));
     EnableResult = false;
     return Result;
   }