+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added --disable-mlp flag to turn MLP into complex material where possible
 - Removed weapons/armor from dynamic cubemap blocklist
 - Added --high-mem option for paster processing in exchange for high memory usage
+- Added wide string support in filenames
 
 ## [0.5.7] - 2025-09-05